Simple Scaling workaround.  +Cleanup
Implemented a way to scale up both sprites and our background image using pygame's built in scale2x() function for surfaces.  This is to be seen as a workaround and not final.  Due to this, boundaries have been set according to the scaling.  However they don't seem to be 2:1.

Also cleaned the logic function housed in our map init.  Moved all python files into a respective directory.  test.py is now named main.py.  main.py is found in the top level directory, unlike the others.